MSBUILD is not DEVENV and debug temp

+++ src/build.sh
@@ -10,7 +10,7 @@ cmake --build ./
 if [[ $RUNNER_OS == "Linux" ]]; then
 	make
 elif [[ $RUNNER_OS == "Windows" ]]; then
-	eval "$DEVENV ./$PROJECT_NAME.sln \"/property:Configuration=Release\""
+	eval "$MSBUILD ./$PROJECT_NAME.sln \"/property:Configuration=Release\""
 else
 	echo "Unsupported OS: $RUNNER_OS"
 	exit 1
@@ -30,14 +30,12 @@ fi
 if [ -d ../build/$RUNNER_OS-$ARCH/ ]; then
 	rm -r -f ../build/$RUNNER_OS-$ARCH/
 fi
-echo $(ls)
-echo $(ls ./x64)
-echo $(ls ./x64/Release)
-echo $(ls ./Release)
 mkdir -p ../build/$RUNNER_OS-$ARCH/
 [[ -f ./main ]] && cp -f ./main ../build/$RUNNER_OS-$ARCH/$PROJECT_NAME
 [[ -f ./Release/main.exe ]] && cp -f ./Release/main.exe ../build/$RUNNER_OS-$ARCH/$PROJECT_NAME.exe
 
 # clean up temporary directory
 cd ../
-rm -r -f ./temp
+if [[ $RUNNER_OS == "Linux" ]]; then
+	rm -r -f ./temp
+fi
